requirements.txt
Flask==2.3.3
Pillow==10.2.0
